### 2021-04-09 #3157 @kthyng
Hi @jhollist! Would you be interested in reviewing this submission for JOSS? We are asking for reviews to be completed within about 4-6 weeks but would suggest you start earlier than that for this iterative process. As a refresher, you can see reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
