### 2021-04-12 #3157 @jhollist
Disregarded!  Good luck with the review.

On Fri, Apr 9, 2021 at 12:20 PM Kristen Thyng ***@***.***>
wrote:

> @jhollist <https://github.com/jhollist> eep! I just saw that your
> preferred language is R and this package is in Python. Sorry for the
> confusion and please disregard this invitation!
>
