### 2021-04-26 #3157 @kthyng
Hi @cheginit I know you did a JOSS review pretty recently so feel free to say no, but are you interested in reviewing this submission for JOSS? As you may know, we would hope for reviews to be finished in 4–6 weeks but with the process starting earlier than that since it is an iterative process.
