### 2021-04-26 #3157 @kthyng
@JannisHoch are you interested in reviewing this JOSS submission? We would hope for reviews to be finished in 4–6 weeks but with the process starting earlier than that since it is an iterative process. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
