### 2021-04-27 #3157 @kthyng
Thanks @JannisHoch! And @cheginit do you think this relationship would impact your ability to be impartial in your review?
