### 2021-04-28 #3221 @whedon
[REVIEW]: mosartwmpy: A Python implementation of the MOSART-WM coupled hydrologic routing and water management model
**Submitting author:** @thurber (<a href="http://orcid.org/0000-0002-4370-9971">Travis Thurber</a>)
**Repository:** <a href="https://github.com/IMMM-SFA/mosartwmpy" target ="_blank">https://github.com/IMMM-SFA/mosartwmpy</a>
**Version:** v0.0.8
