### 2021-05-06 af50ca8
adds unit tests for input/output methods; closes #37
