### 2021-05-06 0b8f6d7
try to fix windows build by installing slugify
