### 2021-05-06 a3b9f1a
try to fix windows build by installing fsutil
