### 2021-05-06 17e7e49
set class vars for package data
