### 2021-05-06 ef1d423
remove pre rasterio build
