### 2021-05-27 f73a5be
update paper based on feedback from #39
