+++ paper/paper.bib
@@ -31,6 +31,35 @@
   month = {apr}
 }
 
+@misc{statemod,
+  title = {StateMod-15.00.14},
+  author = { Colorado's Decision Support Systems (CDSS)},
+  url = {https://github.com/OpenCDSS/cdss-app-statemod-fortran},
+  howpublished = {[Computer Software]},
+  year = {2018},
+  month = {jun}
+}
+
+@misc{glofrim,
+  title = {GLOFRIM v2.1 (Version v2.1)},
+  author = {Jannis Hoch, Dirk Eilander, Hiroaki Ikeuchi, Hessel Winsemius, & Fedor Baart},
+  url = {https://github.com/openearth/glofrim/tree/v2.1},
+  doi = {10.5281/zenodo.4304656},
+  howpublished = {[Computer Software]},
+  year = {2020},
+  month = {dec}
+}
+
+@misc{modflow,
+  title = {MODFLOW 6 Modular Hydrologic Model version 6.2.1},
+  author = {Langevin, C.D., Hughes, J.D., Banta, E.R., Provost, A.M., Niswonger, R.G., and Panday, Sorab},
+  url = {https://doi.org/10.5066/F76Q1VQV},
+  doi = {10.5066/F76Q1VQV},
+  howpublished = {[Computer Software]},
+  year = {2021},
+  month = {feb}
+}
+
 @article{peckham2013component,
   title = {A component-based approach to integrated modeling in the geosciences: The design of CSDMS},
   author = {Peckham, Scott D and Hutton, Eric WH and Norris, Boyana},
@@ -118,20 +147,6 @@
   doi = {10.5194/hess-24-1275-2020}
 }
 
-@article{voisin2017conus,
-  author = {Voisin, Nathalie and Hejazi, Mohamad I. and Leung, L. Ruby and Liu, Lu and Huang, Maoyi and Li, Hong-Yi and Tesfa, Teklu},
-  title = {Effects of spatially distributed sectoral water management on the redistribution of water resources in an integrated water model},
-  journal = {Water Resources Research},
-  volume = {53},
-  number = {5},
-  pages = {4253-4270},
-  keywords = {earth system model, integrated assessment model, withdrawals and consumptive use, sectoral water management, groundwater/surface water allocation},
-  doi = {10.1002/2016WR019767},
-  url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/2016WR019767},
-  eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/2016WR019767},
-  year = {2017}
-}
-
 @Article{hess-19-3239-2015,
   AUTHOR = {Chaney, N. W. and Herman, J. D. and Reed, P. M. and Wood, E. F.},
   TITLE = {Flood and drought hydrologic monitoring: the role of model parameter uncertainty},
@@ -142,4 +157,19 @@
   PAGES = {3239--3251},
   URL = {https://hess.copernicus.org/articles/19/3239/2015/},
   DOI = {10.5194/hess-19-3239-2015}
+}
+
+@article{https://doi.org/10.1029/2018MS001583,
[…]
