+++ paper/paper.md
@@ -43,7 +43,7 @@ MOSART-WM is often utilized as the hydrological component in a larger suite of e
 The target audience for `mosartwmpy` is the data scientist or hydrologist who wishes to rapidly prototype, test, and develop novel modeling capabilities relating to reservoirs and water demand. `mosartwmpy` can be accessed on GitHub (https://github.com/IMMM-SFA/mosartwmpy), and a walkthrough of key functionality and use can be found here: [tutorial](https://mosartwmpy.readthedocs.io/en/latest/). Model results have been validated against historical simulations [@https://doi.org/10.1029/2020WR027902], and a validation utility is included with the code.
 
 # State of the field
-In addition to `mosartwmpy`'s ancestor MOSART-WM, several other models are commonly used in hydrologic modeling, each excelling at different processes. The Community Land Model (CLM) [@https://doi.org/10.1029/2018MS001583] focuses on the traditional water cycle (i.e. precipitation, evaporation) as well as plant and soil chemistry; runoff output from CLM can be used as input for `mosartwmpy`. StateMod [@statemod] focuses on water allocation based on legal constraints (water rights) as well as supply and demand. MODFLOW [@modflow] focuses on solving for complex groundwater flow properties on three-dimensional grids. GLOFRIM [@glofrim] focuses on providing a framework to couple models implementing the BMI across multi-scale grids, for instance by coupling a meteorological model to a water cycle model to river routing model to hydrodynamic model. In this context, `mosartwmpy` focuses on the interactions between river routing, reservoir management, and water demand.
+In addition to `mosartwmpy`'s ancestor MOSART-WM, several other models are commonly used in hydrologic modeling, each excelling at different processes. The Community Land Model [CLM; @https://doi.org/10.1029/2018MS001583] focuses on the traditional water cycle (i.e. precipitation, evaporation) as well as plant and soil chemistry; runoff output from CLM can be used as input for `mosartwmpy`. StateMod [@statemod] focuses on water allocation based on legal constraints (water rights) as well as supply and demand. MODFLOW [@modflow] focuses on solving for complex groundwater flow properties on three-dimensional grids. GLOFRIM [@hoch2017glofrim] focuses on providing a framework to couple models implementing the BMI across multi-scale grids, for instance by coupling a meteorological model to a water cycle model to river routing model to hydrodynamic model. In this context, `mosartwmpy` focuses on the interactions between river routing, reservoir management, and water demand.
 
 # Functionality
 Model input for `mosartwmpy` consists of channel and reservoir geometry, groundwater and subsurface runoff (i.e. rain and ice melt), and water demand. The runoff populates the local streams and river channels, which generally flow toward the oceans (as in \autoref{fig:flow}). In many locations along the main channels, reservoirs collect and store water, releasing portions of the storage downstream over time for various societal uses (i.e. hydroelectricity, drinking water, flood control, and irrigation). Water is also consumed directly from the channels and tributaries for irrigation or non-irrigation uses.
