+++ paper/paper.bib
@@ -33,21 +33,22 @@
 
 @misc{statemod,
   title = {StateMod-15.00.14},
-  author = { Colorado's Decision Support Systems (CDSS)},
+  author = {CDSS},
   url = {https://github.com/OpenCDSS/cdss-app-statemod-fortran},
   howpublished = {[Computer Software]},
   year = {2018},
   month = {jun}
 }
 
-@misc{glofrim,
-  title = {GLOFRIM v2.1 (Version v2.1)},
-  author = {Jannis Hoch, Dirk Eilander, Hiroaki Ikeuchi, Hessel Winsemius, & Fedor Baart},
-  url = {https://github.com/openearth/glofrim/tree/v2.1},
-  doi = {10.5281/zenodo.4304656},
-  howpublished = {[Computer Software]},
-  year = {2020},
-  month = {dec}
+@article{hoch2017glofrim,
+  title={GLOFRIM v1. 0--A globally applicable computational framework for integrated hydrological--hydrodynamic modelling},
+  author={Hoch, Jannis M and Neal, Jeffrey C and Baart, Fedor and Beek, Rens van and Winsemius, Hessel C and Bates, Paul D and Bierkens, Marc FP},
+  journal={Geoscientific Model Development},
+  volume={10},
+  number={10},
+  pages={3913--3929},
+  year={2017},
+  publisher={Copernicus GmbH}
 }
 
 @misc{modflow,
