### 2021-05-31 IMMM-SFA/mosartwmpy#41 @cheginit
[JOSS review] Logging and verbosity
Since many Python users use Jupyter as their IDE, the logging info should be redirected to `stdout` so the logs don't show up with red background. You can do so by configuring the logger, for example, as follows:

```python
import logging
import sys

logger = logging.getLogger(__name__)
logger.setLevel(logging.INFO)
handler = logging.StreamHandler(sys.stdout)
handler.setFormatter(logging.Formatter(""))
logger.handlers = [handler]
logger.propagate = False
```

Also, the model is unnecessarily verbose by default. It might not matter much when you run the code via command-line but when you run it via Jupyter it causes issues. A much better approach is to use `tqdm`. I think a quick fix can be reducing the verbosity by default.

EDIT: Reference to openjournals/joss-reviews#3221
