### 2021-05-31 IMMM-SFA/mosartwmpy#42 @cheginit
[JOSS review] Hard-coded output directory
I think [hard-coding](https://github.com/IMMM-SFA/mosartwmpy/blob/8325dabe7cb4a58281834b3f56492c11310541a8/mosartwmpy/output/output.py#L57) the output directory can cause unintended issues since the outputs are usually large files. The user might prefer to save the outputs to specific directories. Either you should explicitly mention that that the output directory is hard-coded and cannot be changed or make it configurable (my personal preference).

The README file says:
> By default, key model variables are output on a monthly basis at a daily averaged resolution to `./output/<simulation name>/<simulation name>_<year>_<month>.nc`. See the configuration file for examples of how to modify the outputs, and the `./mosartwmpy/state/state.py` file for state variable names.

This doesn't suggest that the output directory is hard-coded and suggests that it can be changed via the `config` file.

EDIT: Reference to openjournals/joss-reviews#3221
