### 2021-05-31 IMMM-SFA/mosartwmpy#43 @cheginit
[JOSS review] Python version and Dependecies
I am wondering what's reason behind not supporting Python 3.7+. What specific features of Python 3.9 is being used in the code? Pinning the dependencies can limit the usage of the project with other packages. Although one can create several environments, having such unnecessary pinned dependencies might affects the user experience.

The `requirements.txt` includes unnecessary dependencies:

```bash
recommonmark
setuptools
sphinx
sphinx-rtd-theme
```

Some other suggestions:

* I suggest to use `conda` instead of `pyenv`, since you're relying on scientific Python libraries, they tend to be orders of magnitudes faster when installed via `conda-forge`.
* You don't think you need both `pyarrow` and `fastparquet` since they both do the same thing. `pyarrow` is the default `parquet` engine in `pandas`.
* When using `h5netcdf` instead of `netcdf4` you need to be aware of, and inform the user about, the difference that are documented on [`xarray`'s website](https://xarray.pydata.org/en/stable/user-guide/io.html):
> There may be minor differences in the Dataset object returned when reading a NetCDF file with different engines. For example, single-valued attributes are returned as scalars by the default engine=netcdf4, but as arrays of size (1,) when reading with engine=h5netcdf.

Overall, I think the dependencies need a reevaluation.

EDIT: Reference to openjournals/joss-reviews#3221
