### 2021-05-31 #3221 @cheginit
Overall, I find this submission a useful contribution to the hydrology community and appreciate the effort of the developers. Here are my initial thoughts on the submission.

**Paper**

Overall, the paper reads well and states the purpose of the project clearly. Some of my concerns have already been addressed by the new commits. However, I am not quite sure if I understand the purpose of this sentence:
> As a sample benchmark, a one-year simulation using MOSART-WM on a 24-CPU node of an institutional supercomputer can be performed in about half an hour; whereas the same simulation using `mosartwmpy` on a personal 6-CPU laptop can be performed in about five hours.

This is not a fair and informative comparison: Running a Fortran code on a cluster with 24 CPUs vs. running a Python code on a latop with 6 CPUs. There's no common ground between the two; Code languages are different, system specifications of the testing environments are different, and number of CPUs are different. I know setting up E3SM on a laptop is not an easy task, so I would suggest to at least run the Python code on the same cluster with 8 CPUs and run the Fortran version on the same cluster and with 8 CPUs. Overall, I am not sure if this comparison is really necessary since, as you mentioned in the paper, the main purpose of `mosartwmpy` is prototyping and educational.

**Code**

The authors opted to use `numexpr` as the engine, so to speak, to carry out the heavy lifting in their implementation of MOSART-WM. This choice makes the code difficult to read and hinders further development by the "the domain scientist". There are much better approaches that even scale much better (more than 8 CPUs) than `numexpr`. I would like to know why the developers decided to use `numexpr` over [`numba`](https://numba.pydata.org/)?

I understand that this might be a personal preference but I believe in the future versions, the authors might want to consider refactoring their code to use `numba` as the engine. It can improve the performance and scaling, as well as, readability. I developed a hydrological model (not public yet) that is based on `numba` and have tested it on a cluster with 28 cores and it scales nicely.

Regarding the documentation, the website provides examples but I think it can be improved by providing an example for working with restart files as well as runoff outputs from CLM or E3SM outputs since this is the primary use case.

Additionally, I opened a couple of issues/suggestions in the repository.
