### 2021-05-31 IMMM-SFA/mosartwmpy#41 @cheginit
Here's an example screenshot from Jupyter:
![image](https://user-images.githubusercontent.com/13016644/120224913-f07ab580-c209-11eb-9150-9f3c95c71da8.png)
