+++ .github/workflows/build_windows.yml
@@ -26,12 +26,12 @@ jobs:
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/numpy-1.20.2+mkl-cp39-cp39-win_amd64.whl
-            pip install numpy-1.20.2+mkl-cp39-cp39-win_amd64.whl
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/GDAL-3.2.2-cp39-cp39-win_amd64.whl
-            pip install GDAL-3.2.2-cp39-cp39-win_amd64.whl
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/rasterio-1.2.3-cp39-cp39-win_amd64.whl
-            pip install rasterio-1.2.3-cp39-cp39-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/numpy-1.20.3+mkl-cp39-cp39-win_amd64.whl
+            pip install numpy-1.20.3+mkl-cp39-cp39-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/GDAL-3.3.0-cp39-cp39-win_amd64.whl
+            pip install GDAL-3.3.0-cp39-cp39-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/rasterio-1.2.4-cp39-cp39-win_amd64.whl
+            pip install rasterio-1.2.4-cp39-cp39-win_amd64.whl
             pip install -r requirements.txt
             python setup.py install
             pip install numpy --upgrade
@@ -40,5 +40,4 @@ jobs:
         - name: Test and generate coverage report
           run: |
             pip install pytest
-            pip install pytest-cov
-            pytest --cov=./ --cov-report=xml
+            pytest
