### 2021-06-01 IMMM-SFA/mosartwmpy#41 @thurber
This is a good point, I can adjust the majority of the logging to be debug level, and only echo the configuration into the file logger.

Curious, from your perspective, is logging the current timestep useful? Perhaps just at daily or monthly or yearly intervals?
