### 2021-06-01 IMMM-SFA/mosartwmpy#42 @thurber
Very true, I left a note for myself in #31 to fix this, but hadn't done it yet. I will make the output directory configurable.
