### 2021-06-01 IMMM-SFA/mosartwmpy#43 @thurber
Fair point, many folks are bringing up the Python version concerns. I briefly tried to make use of some of the new shared memory semantics of 3.9, but abandoned that effort -- honestly my main concern is to push folks to keep up to date on their software tooling, but that isn't really the job of a library. I will update to ensure support for 3.7+
