+++ .github/workflows/build_windows.yml
@@ -1,6 +1,7 @@
 name: windows
 
-on: [push]
+#on: [push]
+on: workflow_dispatch
 
 jobs:
   build:
