### 2021-06-01 #3221 @thurber
Thanks @cheginit, this is good feedback!

I just merged a PR that removes most of the _Functionality and limitations_ section based on @JannisHoch's feedback, so the section with the awkward CPU comparison is gone.

Regarding the speed and the use of `numexpr` -- I agree it's semantically burdensome, but at least for my own laptop and cluster, the `numexpr` implementation was faster than a `numba` implementation for doing the same vectorized calculations. That said, I'm still not satisfied with the execution speed and am planning to do an overhaul in the medium-term future. I'd be very curious to learn from your experiences with `numba`. In your implementation, do you solve in a vectorized fashion or cell-by-cell? Despite the traditional advice to go vectorized in Python, I fear it may have led to the current speed challenges we are facing in `mosartwmpy`.

Good idea about providing examples with restart files and dynamic runoff inputs -- I'll open an issue for that and get to work on it.
