### 2021-06-01 IMMM-SFA/mosartwmpy#41 @cheginit
I think if you go with `tqdm` daily is fine. Otherwise, I think monthly or yearly works better as you said. I wonder if you tested the impact of logging on the performance? Since IO can sometimes affect the performance significantly, depending on the time each time-step takes.
