### 2021-06-01 IMMM-SFA/mosartwmpy#43 @cheginit
Yes, I understand. The main issue with not supporting the Python versions that haven't reached EOL is compatibility with other libraries. Version 3.6's EOL is in 7 months and most of the major libraries have already dropped support for it. So I think 3.7+ support is reasonable.
