### 2021-06-01 #3221 @cheginit
@thurber The recommended way of implementing `numba` is to avoid vectorization and unroll all the loops. Adding signatures to functions can also speed up the code. You need to keep type casting during the runtime at the minimum. Another point is about the decision between using `f8` or `f4` that also affects the performance and accuracy.
