### 2021-06-03 a8d220c
add "simulation.output_path" configuration option; fixes #31 and #42
