### 2021-06-03 721b5d5
reduces logging verbosity at INFO level; fixes #41
