+++ mosartwmpy/output/output.py
@@ -12,7 +12,7 @@ from mosartwmpy.utilities.timing import timing
 def initialize_output(self):
     """Initializes the output buffer."""
     
-    logging.info('Initalizing output buffer.')
+    logging.debug('Initalizing output buffer.')
     if self.config.get('simulation.output_resolution') % self.config.get('simulation.timestep') != 0 or self.config.get('simulation.output_resolution') < self.config.get('simulation.timestep'):
         raise Exception('The `simulation.output_resolution` must be greater than or equal to and evenly divisible by the `simulation.timestep`.')
     for output in self.config.get('simulation.output'):
@@ -102,7 +102,7 @@ def write_output(self):
                 frame[output.get('name')].attrs['units'] = output.get('units')
 
     # if file exists and it's not a new period, update existing file else write to new file and include grid variables
-    logging.info(f'Writing to output file: {Path(filename)}.')
+    logging.debug(f'Writing to output file: {Path(filename)}.')
     if not is_new_period and Path(filename).is_file():
         nc = open_dataset(Path(filename)).load()
         # slice the existing data to account for restarts
@@ -162,5 +162,5 @@ def write_restart(self):
     x = self.state.to_dataframe().to_xarray()
     filename = Path(f'{self.config.get("simulation.output_path")}/{self.name}/restart_files/{self.name}_restart_{self.current_time.year}_{self.current_time.strftime("%m")}_{self.current_time.strftime("%d")}.nc')
     x = x.rio.write_crs(4326)
-    logging.info(f'Writing restart file: {filename}.')
+    logging.debug(f'Writing restart file: {filename}.')
     x.to_netcdf(filename)
