### 2021-06-03 a016886
lower python requirement to 3.7; move dev dependencies into extras; fixes #43
