+++ requirements.txt
@@ -13,12 +13,8 @@ pathvalidate==2.3.0
 psutil==5.7.3
 pyarrow==3.0.0
 python-benedict==0.24.0
-recommonmark==0.7.1
 regex==2020.11.13
 requests==2.25.1
 rioxarray==0.3.1
-setuptools==54.1.2
-sphinx==3.5.2
-sphinx-rtd-theme==0.5.1
 tqdm==4.60.0
 xarray==0.16.1
