### 2021-06-03 IMMM-SFA/mosartwmpy#43 @thurber
In #46, I have set the minimum python version to 3.7 and moved the dependencies related to building the documentation into the `extra_require` section such that they are optional. Also, removed `fastparquet` dependency since I'm not using it anywhere.

Wanted to respond to your other points as well:
* conda -- I think this is a good idea; I need to learn how to submit a package to conda-forge, so I will make a separate issue for this
* `h5netcdf` vs `netcdf4` -- for all the model outputs i'm using the default `netcdf4` engine; the one place `h5netcdf` is used is when creating a compressed file from all the static grid/model variables, which is mostly just used for the unit tests (i found this engine created slightly smaller files and read them slightly faster). It is unnecessary complexity though so I will make another issue to back out this requirement.
