### 2021-06-04 IMMM-SFA/mosartwmpy#49 @JannisHoch
JOSS-REVIEW: documentation
hi @thurber,

i find the documentation overall good, but also pretty concise.
it would be great to include a jupyter notebook with an example application of your model.
this would not only give users an idea how a typical model workflow would look like, but also provide you with an opportunity to explain the model's functionality and set-up step-by-step and in greater detail.

i furthermore would like to see a 'statement of need' (similar as in the paper) in the documentation. not everyone is going to read the paper, but many will just go through the readme or rtd.

reference to: https://github.com/openjournals/joss-reviews/issues/3221.
