### 2021-06-04 IMMM-SFA/mosartwmpy#49 @JannisHoch
in addition to this:
the example application could be based on example data (if their size allows), which in turn could be used to provide a simple model which users can use to discover and play around with model functionalities.
