### 2021-06-04 IMMM-SFA/mosartwmpy#43 @cheginit
That's looks great!

Regarding conda, that's a (strong) suggestion and not a requirement for JOSS. For development and testing, I highly recommend [MambaForge](https://github.com/conda-forge/miniforge). For Github Actions, you can take a look at the [action file](https://github.com/cheginit/pynhd/blob/main/.github/workflows/test.yml) that I am using in one of my projects. It uses MambaForge and caching.

Regarding `h5netcdf`, I agree that it's an unnecessary complexity.
