### 2021-06-04 IMMM-SFA/mosartwmpy#49 @thurber
good idea, i'll add a jupyter notebook with the tutorial inside it
