### 2021-06-04 bacde88
remove h5netcdf dependency; fixes #48
