+++ requirements.txt
@@ -1,19 +1,18 @@
 bmipy==2.0
-dask[complete]==2021.01.1
+dask[complete]==2021.5.1
 epiweeks==2.1.2
-h5netcdf==0.10.0
-matplotlib==3.3.3
+matplotlib==3.4.2
 nc-time-axis==1.2.0
-netCDF4==1.5.4
-numexpr==2.7.2
-numpy==1.19.4
-pandas==1.1.4
+netCDF4==1.5.6
+numexpr==2.7.3
+numpy==1.20.3
+pandas==1.2.4
 pathvalidate==2.3.0
-psutil==5.7.3
-pyarrow==3.0.0
+psutil==5.8.0
+pyarrow==4.0.1
 python-benedict==0.24.0
-regex==2020.11.13
+regex==2021.4.4
 requests==2.25.1
-rioxarray==0.3.1
+rioxarray==0.4.2
 tqdm==4.60.0
-xarray==0.16.1
+xarray==0.18.2
