### 2021-06-04 IMMM-SFA/mosartwmpy#49 @cheginit
@thurber Since you're working on the documentation, just wanted to remind you to update the installation instructions (using Python 3.7+)
