### 2021-06-04 073b6d2
remove epiweeks requirement for now, in preparation for conda-forge version; bump version
