+++ requirements.txt
@@ -1,6 +1,5 @@
 bmipy==2.0
 dask[complete]==2021.5.1
-epiweeks==2.1.2
 matplotlib==3.4.2
 nc-time-axis==1.2.0
 netCDF4==1.5.6
