### 2021-06-04 5648cb1
move requirements to setup.py
