+++ requirements.txt
@@ -1,17 +1,1 @@
-bmipy==2.0
-dask[complete]==2021.5.1
-matplotlib==3.4.2
-nc-time-axis==1.2.0
-netCDF4==1.5.6
-numexpr==2.7.3
-numpy==1.20.3
-pandas==1.2.4
-pathvalidate==2.3.0
-psutil==5.8.0
-pyarrow==4.0.1
-python-benedict==0.24.0
-regex==2021.4.4
-requests==2.25.1
-rioxarray==0.4.2
-tqdm==4.60.0
-xarray==0.18.2
+.[dev]
