### 2021-06-07 #3221 @cheginit
@thurber @crvernon The submission is almost ready for submission. In my opinion, there are only two things left to be addressed:

- [ ] Updating the installation instructions in the README and the website to reflect the latest changes such as support for Python 3.7+.
- [ ] Adding Example(s) for running a model with outputs from other runoff generation packages such as CESM and E3SM.
