+++ .github/workflows/build_windows.yml
@@ -1,7 +1,6 @@
 name: windows
 
-#on: [push]
-on: workflow_dispatch
+on: [push]
 
 jobs:
   build:
@@ -13,7 +12,7 @@ jobs:
 
       env:
         OS: ${{ matrix.os }}
-        PYTHON: '3.9'
+        PYTHON: '3.7.2'
 
       steps:
 
@@ -22,21 +21,18 @@ jobs:
         - name: Set up Python
           uses: actions/setup-python@master
           with:
-            python-version: 3.9
+            python-version: 3.7.2
 
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/numpy-1.20.3+mkl-cp39-cp39-win_amd64.whl
-            pip install numpy-1.20.3+mkl-cp39-cp39-win_amd64.whl
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/GDAL-3.3.0-cp39-cp39-win_amd64.whl
-            pip install GDAL-3.3.0-cp39-cp39-win_amd64.whl
-            curl -O https://download.lfd.uci.edu/pythonlibs/w4tscw6k/rasterio-1.2.4-cp39-cp39-win_amd64.whl
-            pip install rasterio-1.2.4-cp39-cp39-win_amd64.whl
-            pip install -r requirements.txt
-            python setup.py install
-            pip install numpy --upgrade
-            pip install python-benedict --upgrade
+            curl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl --o np.whl
+            pip install np.whl
+            curl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/GDAL-3.2.3-cp37-cp37m-win_amd64.whl --o gdal.whl
+            pip install gdal.whl
+            curl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/rasterio-1.2.4-cp37-cp37m-win_amd64.whl --o rasterio.whl
+            pip install rasterio.whl
+            pip install .
 
         - name: Test and generate coverage report
           run: |
