+++ .github/workflows/build_osx.yml
@@ -12,7 +12,7 @@ jobs:
 
       env:
         OS: ${{ matrix.os }}
-        PYTHON: '3.9'
+        PYTHON: '3.7.2'
 
       steps:
 
@@ -21,13 +21,12 @@ jobs:
         - name: Set up Python
           uses: actions/setup-python@master
           with:
-            python-version: 3.9
+            python-version: 3.7.2
 
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install -r requirements.txt
-            python setup.py install
+            pip install .
 
         - name: Test and generate coverage report
           run: |
