+++ .github/workflows/build_osx.yml
@@ -12,7 +12,7 @@ jobs:
 
       env:
         OS: ${{ matrix.os }}
-        PYTHON: '3.7.2'
+        PYTHON: '3.7'
 
       steps:
 
@@ -21,7 +21,7 @@ jobs:
         - name: Set up Python
           uses: actions/setup-python@master
           with:
-            python-version: 3.7.2
+            python-version: 3.7
 
         - name: Install dependencies
           run: |
