+++ .github/workflows/build_linux.yml
@@ -28,7 +28,7 @@ jobs:
             python -m pip install --upgrade pip
             pip install .
 
-        - name: Test and generate coverage report
+        - name: Test
           run: |
             pip install pytest
             pytest
