+++ .github/workflows/build_windows.yml
@@ -26,11 +26,11 @@ jobs:
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            curl -O ./np.whl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl
             pip install numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl
-            curl -O ./gdal.whl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/GDAL-3.2.3-cp37-cp37m-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/GDAL-3.2.3-cp37-cp37m-win_amd64.whl
             pip install GDAL-3.2.3-cp37-cp37m-win_amd64.whl
-            curl -O ./rasterio.whl https://download.lfd.uci.edu/pythonlibs/q4trcu4l/rasterio-1.2.4-cp37-cp37m-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/rasterio-1.2.4-cp37-cp37m-win_amd64.whl
             pip install rasterio-1.2.4-cp37-cp37m-win_amd64.whl
             pip install .
 
