+++ .github/workflows/build_windows.yml
@@ -28,8 +28,8 @@ jobs:
             python -m pip install --upgrade pip
             curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl
             pip install numpy-1.20.3+mkl-cp37-cp37m-win_amd64.whl
-            curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/GDAL-3.2.3-cp37-cp37m-win_amd64.whl
-            pip install GDAL-3.2.3-cp37-cp37m-win_amd64.whl
+            curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/GDAL-3.3.0-cp37-cp37m-win_amd64.whl
+            pip install GDAL-3.3.0-cp37-cp37m-win_amd64.whl
             curl -O https://download.lfd.uci.edu/pythonlibs/q4trcu4l/rasterio-1.2.4-cp37-cp37m-win_amd64.whl
             pip install rasterio-1.2.4-cp37-cp37m-win_amd64.whl
             pip install .
