### 2021-06-07 7fc56f1
update docs to reflect python 3.7+; update handling of coupled input variables; add Jupyter notebook tutorial
