### 2021-06-07 IMMM-SFA/mosartwmpy#45 @thurber
I added an example for restart and dynamic runoff to pull request #50
