### 2021-06-07 IMMM-SFA/mosartwmpy#49 @thurber
I added a Jupyter notebook tutorial in pull request #50!
