### 2021-06-10 IMMM-SFA/mosartwmpy#49 @thurber
addressed by #50
