### 2021-06-10 IMMM-SFA/mosartwmpy#45 @thurber
addressed by #50
