### 2021-06-10 #3221 @thurber
thanks @cheginit, I believe I have addressed both these points in the latest PR, which adds a link to a Jupyter notebook with a tutorial that includes steps to use restart files and steps to drive with dynamic input: https://github.com/IMMM-SFA/mosartwmpy/blob/main/notebooks/tutorial.ipynb
