### 2021-06-14 #3221 @kthyng
@cheginit Thanks for your review! Rather than assuming, can I verify the meaning of crossing out the performance checkbox above? What does that mean in your review?
