### 2021-06-14 #3221 @kthyng
> @kthyng I meant that there was no claim to be checked. In the parenthesis, it says to check it off, I thought that's what it meant. If you prefer, I can just check the item.

Ok that is what I figured, but yes please just check it off instead if you please. Thanks.
