### 2021-06-14 IMMM-SFA/mosartwmpy#49 @JannisHoch
@thurber,

documentation in general is improved.
last thing i'd  recommend is to link the notebook you have under `notebooks` to the readthedocs-documentation via the nblink extension. the package to install is nbconvert (https://nbconvert.readthedocs.io/en/latest/) and then link a `ipynb.nblink` file in index.rst file.
the nblink-file should include a relative path to the ipynb-file, looking like that:

```
{
    "path": "../../filename.ipynb"
}
```
