### 2021-06-14 d640319
add channel to conda install instruction; fixes #47
