### 2021-06-14 72c292a
adds jupyter notebook tutorial to readthedocs; fixes #49
