+++ setup.py
@@ -9,7 +9,7 @@ def readme():
 
 setup(
     name='mosartwmpy',
-    version='0.0.6',
+    version='0.0.7',
     packages=find_packages(),
     url='https://github.com/IMMM-SFA/mosartwmpy',
     license='BSD2-Simplified',
@@ -42,6 +42,7 @@ setup(
     extras_require={
         'dev': [
             'build==0.4.0',
+            'nbsphinx==0.8.6',
             'recommonmark==0.7.1',
             'setuptools==54.1.2',
             'sphinx==3.5.2',
