### 2021-06-14 cf874fc
adds ipython to dependencies
