### 2021-06-14 afa033a
only run linux build on push; others on PR
