+++ .github/workflows/build_windows.yml
@@ -1,6 +1,6 @@
 name: windows
 
-on: [push]
+on: [pull_request]
 
 jobs:
   build:
