+++ .github/workflows/build_linux.yml
@@ -28,7 +28,14 @@ jobs:
             python -m pip install --upgrade pip
             pip install .
 
-        - name: Test
+        - name: Test and generate coverage report
           run: |
             pip install pytest
-            pytest
+            pip install pytest-cov
+            pytest --cov=./ --cov-report=xml
+
+        - name: Upload coverage to Codecov
+          uses: codecov/codecov-action@v1
+          with:
+            file: ./coverage.xml
+            fail_ci_if_error: true
