### 2021-06-17 #3221 @kthyng
@thurber  please make an archive of your software at a place like Zenodo, and make sure that the title and author list exactly match your JOSS paper.
