### 2021-06-17 #3221 @kthyng
@thurber Please check out that invalid DOI listing above. Other paper comments:
* Your first reference (for Github) seems to need an author. Can you find something reasonable to put there?
* Yoon et al title has a spacing issue
* Should "Earth" be capitalized in several of the references? You can preserve capitalization with {} around characters, words, or full bibtex lines.
