### 2021-06-17 92f694e
begin migration from numexpr to numba
