+++ mosartwmpy/subnetwork/state.py
@@ -1,81 +1,45 @@
-import numpy as np
-import numexpr as ne
-
-from mosartwmpy.config.parameters import Parameters
-from mosartwmpy.grid.grid import Grid
-from mosartwmpy.state.state import State
-
-def update_subnetwork_state(state: State, grid: Grid, parameters: Parameters, base_condition: np.ndarray) -> None:
-    """Updates the physical properties of the subnetwork river channels based on current state.
-
-    Args:
-        state (State): the current model state; will be mutated
-        grid (Grid): the model grid
-        parameters (Parameters): the model parameters
-        base_condition (np.ndarray): a boolean array representing where the update should occur in the state
-    """
-        
-    # update state variables
-    length_condition = calculate_length_condition(grid.subnetwork_length, state.subnetwork_storage)
-    state.subnetwork_cross_section_area = calculate_subnetwork_cross_section_area(base_condition, length_condition, state.subnetwork_storage, grid.subnetwork_length, state.subnetwork_cross_section_area)
-    state.subnetwork_depth = calculate_subnetwork_depth(base_condition, length_condition, state.subnetwork_cross_section_area, parameters.tiny_value, grid.subnetwork_width, state.subnetwork_depth)
-    state.subnetwork_wetness_perimeter = calculate_subnetwork_wetness_perimeter(base_condition, length_condition, state.subnetwork_depth, parameters.tiny_value, grid.subnetwork_width, state.subnetwork_wetness_perimeter)
-    state.subnetwork_hydraulic_radii = calculate_subnetwork_hydraulic_radii(base_condition, length_condition, state.subnetwork_wetness_perimeter, parameters.tiny_value, state.subnetwork_cross_section_area, state.subnetwork_hydraulic_radii)
+import numba as nb
 
 
-calculate_length_condition = ne.NumExpr(
-    '(subnetwork_length > 0) & (subnetwork_storage > 0)',
-    (('subnetwork_length', np.float64), ('subnetwork_storage', np.float64))
+@nb.jit(
+    "void("
+        "int64, float64[:], float64[:], float64[:], float64[:],"
+        "float64[:], float64[:], float64[:], float64"
+    ")",
+    nopython=True,
+    nogil=True
 )
+def update_subnetwork_state(
+    i,
+    subnetwork_length,
+    subnetwork_width,
+    subnetwork_storage,
+    subnetwork_cross_section_area,
+    subnetwork_depth,
+    subnetwork_wetness_perimeter,
+    subnetwork_hydraulic_radii,
+    tiny_value,
+):
+    """Updates the physical properties of the subnetwork river channels based on current state."""
 
-calculate_subnetwork_cross_section_area = ne.NumExpr(
-    'where('
-        'base_condition,'
-        'where('
-            'length_condition,'
-            'subnetwork_storage / subnetwork_length,'
-            '0'
-        '),'
-        'subnetwork_cross_section_area'
-    ')',
-    (('base_condition', np.bool), ('length_condition', np.bool), ('subnetwork_storage', np.float64), ('subnetwork_length', np.float64), ('subnetwork_cross_section_area', np.float64))
-)
+    has_water = (subnetwork_length[i] > 0.0) and (subnetwork_storage[i] > 0.0)
 
-calculate_subnetwork_depth = ne.NumExpr(
-    'where('
-        'base_condition,'
-        'where('
-            'length_condition & (subnetwork_cross_section_area > tiny_value),'
-            'subnetwork_cross_section_area / subnetwork_width,'
-            '0'
-        '),'
[…]
