+++ mosartwmpy/update/update.py
@@ -1,6 +1,6 @@
+import numba as nb
 import numpy as np
 import pandas as pd
-import logging
 
 from benedict.dicts import benedict as Benedict
 
@@ -8,20 +8,14 @@ from mosartwmpy.config.parameters import Parameters
 from mosartwmpy.grid.grid import Grid
 from mosartwmpy.state.state import State
 
-from mosartwmpy.direct_to_ocean.direct_to_ocean import direct_to_ocean
-from mosartwmpy.flood.flood import flood
 from mosartwmpy.hillslope.routing import hillslope_routing
-from mosartwmpy.main_channel.irrigation import  main_channel_irrigation
+from mosartwmpy.main_channel.irrigation import main_channel_irrigation
 from mosartwmpy.main_channel.routing import main_channel_routing
 from mosartwmpy.reservoirs.regulation import extraction_regulated_flow, regulation
 from mosartwmpy.subnetwork.irrigation import subnetwork_irrigation
 from mosartwmpy.subnetwork.routing import subnetwork_routing
 
-# ignore numpy NaN and invalid warnings
-# (i.e. divide by zero and NaN logicals -- in Pandas/Dask, these simply remain NaN instead of becoming infinite)
-np.seterr(all='ignore')
-# filter pandas chained assignment warnings -- pretty sure handling this correctly
-pd.options.mode.chained_assignment = None
+from mosartwmpy.utilities.timing import Timer
 
 
 def update(state: State, grid: Grid, parameters: Parameters, config: Benedict) -> None:
@@ -33,202 +27,642 @@ def update(state: State, grid: Grid, parameters: Parameters, config: Benedict) -
         parameters (Parameters): the model parameters
         config (Benedict): the model configuration
     """
-    # perform one timestep
-    
-    # ignore nan, overflow, underflow, and div by 0 warnings, since they are handled correctly
-    np.seterr(all='ignore')
-    
-    ###
-    ### Reset certain state variables
-    ###
-    state.flow[:] = 0
-    state.outflow_downstream_previous_timestep[:] = 0
-    state.outflow_downstream_current_timestep[:] = 0
-    state.outflow_before_regulation[:] = 0
-    state.outflow_after_regulation[:] = 0
-    state.outflow_sum_upstream_average[:] = 0
-    state.lateral_flow_hillslope_average[:] = 0
-    state.runoff[:] = 0
-    state.direct[:] = 0
-    state.flood[:] = 0
-    state.runoff_land[:] = 0
-    state.runoff_ocean[:] = 0
-    state.delta_storage[:] = 0
-    state.delta_storage_land[:] = 0
-    state.delta_storage_ocean[:] = 0
-    if config.get('water_management.enabled', False):
-        state.grid_cell_deficit[:] = 0
 
-    # flood
-    flood(state, grid, parameters, config)
+    # size of flattened grid
+    n = len(grid.id)
 
-    # direct to ocean
-    direct_to_ocean(state, grid, parameters, config)
-    
-    ###
-    ### Subcycling
-    ###
-    
-    # convert runoff to m/s
-    state.hillslope_surface_runoff = state.hillslope_surface_runoff / grid.area
-    state.hillslope_subsurface_runoff = state.hillslope_subsurface_runoff / grid.area
-    state.hillslope_wetland_runoff = state.hillslope_wetland_runoff / grid.area
-
-    # subcycle timestep
-    delta_t =  config.get('simulation.timestep') / config.get('simulation.subcycles')
-    
+    _prepare(
+        n,
+        config.get('water_management.enabled', False),
+        config.get('simulation.timestep'),
+        grid.land_mask,
+        grid.mosart_mask,
+        grid.area,
+        state.flow,
+        state.direct_to_ocean,
[…]
