### 2021-06-24 #3221 @kthyng
@thurber Your changes look good! Does look like that is a valid missing DOI above — please add that to your Hoch reference and we should be good to go!
