### 2021-06-24 961fc4b
add DOI to Hoch reference
