+++ paper/paper.bib
@@ -48,7 +48,8 @@
   number={10},
   pages={3913--3929},
   year={2017},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/gmd-10-3913-2017}
 }
 
 @misc{modflow,
